fix for bug #3459
git-svn-id: svn://source.pentaho.org/svnkettleroot/mirror/Kettle/trunk@1784 5fb7f6ec-07c1-534a-b4ca-9155e429e800

+++ src/be/ibridge/kettle/core/value/ValueInteger.java
@@ -151,7 +151,7 @@ public class ValueInteger implements ValueInterface, Cloneable
 
     public BigDecimal getBigNumber()
     {
-        return new BigDecimal(number);
+        return new BigDecimal((double)number);
     }
 
     public void setBigNumber(BigDecimal number)
